added workflows and tests
