+++ .github/workflows/test-plugin-integration.yml
@@ -1,0 +1,35 @@
+name: Run Tests with Docker
+
+on:
+  push:
+    branches:
+      - main
+  pull_request:
+
+jobs:
+  test:
+    runs-on: ubuntu-latest
+
+    steps:
+      # Checkout the repository
+      - name: Checkout code
+        uses: actions/checkout@v3
+
+      # Modify ADDITIONAL_PLUGS in .local.env
+      - name: Update ADDITIONAL_PLUGS in .local.env
+        run: |
+          sed -i '/^ADDITIONAL_PLUGS=/c\ADDITIONAL_PLUGS=[{\"name\": \"camera\", \"package_name\": \"git+https://github.com/ohcnetwork/care_camera_asset.git\", \"version\": \"@main\", \"configs\": {}}]' ./docker/.local.env
+
+      # Build and start containers
+      - name: Build and start Docker containers
+        run: |
+          make up docker_config_file='docker-compose.local.yaml'
+
+      # Run tests using `make test`
+      - name: Run `make test`
+        run: make test
+
+      # Run Django management command tests
+      - name: Run Django management command tests
+        run: |
+          docker compose exec backend bash -c "python manage.py test camera --keepdb --parallel --shuffle"
